+++ piwigo/Upload/Pick Pasteboard Images/PasteboardImagesViewController.swift
@@ -956,7 +956,7 @@ class PasteboardImagesViewController: UIViewController, UICollectionViewDataSour
                 if let resizeImageOnUpload = uploadParameters["resizeImageOnUpload"] as? Bool {
                     updatedRequest.resizeImageOnUpload = resizeImageOnUpload
                     if resizeImageOnUpload {
-                        if let photoResize = uploadParameters["photoResize"] as? Int {
+                        if let photoResize = uploadParameters["photoResize"] as? Int16 {
                             updatedRequest.photoResize = photoResize
                         }
                     } else {
@@ -966,7 +966,7 @@ class PasteboardImagesViewController: UIViewController, UICollectionViewDataSour
                 if let compressImageOnUpload = uploadParameters["compressImageOnUpload"] as? Bool {
                     updatedRequest.compressImageOnUpload = compressImageOnUpload
                 }
-                if let photoQuality = uploadParameters["photoQuality"] as? Int {
+                if let photoQuality = uploadParameters["photoQuality"] as? Int16 {
                     updatedRequest.photoQuality = photoQuality
                 }
                 if let prefixFileNameBeforeUpload = uploadParameters["prefixFileNameBeforeUpload"] as? Bool {
